kill the deadlock in database
make sure, the transactions has finished when the request has finished.

Hopefully, one of the longest hunt for a deadlock in my life is now over
:)

Problem was, that even a db_session.query... causes the start of
a new session putting a read lock on the database.

So any write processes afterwards wait for the original transaction to
be closed, which might not happen for a long time.